# [opt](func) opt the revert_null_map performance