## Proposed changes

```
2024-05-04T23:54:46+08:00
Running ./benchmark
Run on (96 X 3100.01 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 36608 KiB (x2)
Load Average: 0.31, 1.20, 1.20
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
origin_test   10682025 ns     10682054 ns           55
simd_test       966283 ns       966286 ns          660
```

<!--Describe your changes.-->

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...